pyproject.toml
[project]
name = "tech-challenge-response"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic-ai-slim[google]>=0.8.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.24.0",
]
dev = [
    "pre-commit>=3.6.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "bandit>=1.7.5",
]
